requirements.txt
streamlit
pandas
openpyxl
plotly
requests
numpy
streamlit-autorefresh
sqlalchemy
psycopg2-binary
python-dotenv
